Include support for generating new trampoline types as unsupported API.
This provides a trampoline_table.h containing the necessary API to
manage the allocation of trampoline tables. This API is unsupported,
and may be included directly by first defining PL_BLOCKIMP_PRIVATE.

Additionally, the gentramp.sh script is included as a framework
resource. It may be used to generate arbitrary trampoline tables using
the .tramp file format as used by the imp_implementationWithBlock
example implementation. This too is considered unsupported API, and may
be removed or break at any time.